Carbon-Epoxy Plate with dimensions
Considering a carbon - epoxy plate with width and length of 0.3 m and 0.3 m, with a thickness of 0.001 m. This plate is under an axial force of 80 kN in direction of the fiber; and also it's under compression of 5 kN in perpendicular direction to the fiber. 
We need to determine the change on the dimensions on the plane of the plate, considering the following mechanical properties for the carbon - epoxy system.